+++ changelog.md
@@ -1,3 +1,13 @@
+## Version 0.1.3
+
+### New Features
+- Added Support for the binary format docx in the from section (Manual Converter only (Needs Pandoc 1.13.1 or better)
+- Added Preferences for the Manual Converter
+- Added Support for error-messages from the Manual Converter
+
+### Bug Fixes
+- Binary Handling of the Manual Converter (For odt, docx formats)
+
 ## Version 0.1.2
 
 ### Bug Fixes